+++ docs-src/docs/api/hooks/demos/data.ts
@@ -1,0 +1,20 @@
+export function dictFactory() {
+  return {
+    name: 'dict',
+    a: {
+      b: { c: 1 },
+      b1: { c1: 1 },
+    },
+    info: { name: 'helux', age: 1 },
+    desc: 'awesome lib',
+    extra: {
+      list: [],
+      mark: 'extra',
+      toBeDrive: 0,
+      prefixedMark: '',
+      newName: '',
+    },
+    f: 1,
+    g: 1,
+  };
+}