+++ packages/helux-core/src/factory/creator/globalId.ts
@@ -18,9 +18,9 @@ export function initGlobalEmpty(apiCtx: CoreApiCtx, createFn: Fn) {
   let shared = ctx.globalEmpty;
   if (!shared) {
     // global shared state
-    const { state } = createFn({ apiCtx, rawState: {}, forGlobal: true, stateType: STATE_TYPE.GLOGAL_EMPTY });
-    const internal = getInternal(state);
-    ctx.globalEmpty = state;
+    const { stateRoot } = createFn({ apiCtx, rawState: {}, forGlobal: true, stateType: STATE_TYPE.GLOGAL_EMPTY });
+    const internal = getInternal(stateRoot);
+    ctx.globalEmpty = stateRoot;
     ctx.globalEmptyInternal = internal;
   }
   GLOBAL_EMPTY = shared;