+++ packages/helux-core/src/factory/creator/parse.ts
@@ -20,7 +20,7 @@ import type {
   MutateFn,
   MutateFnStdDict,
   NumStrSymbol,
-  WatchDepFn,
+  WatchFnDeps,
   WatchOptionsType,
 } from '../../types/base';
 import { genFnKey } from '../common/key';
@@ -310,15 +310,15 @@ export function parseRules(options: ParsedOptions): IRuleConf {
 
 export function parseCreateMutateOpt(descOrOptions?: string | IRunMutateOptions) {
   // 不设定 desc 的话，默认指向可能存在的单函数
-  const { out = true, desc = SINGLE_MUTATE, strict = false } = {};
+  const { desc = SINGLE_MUTATE, strict = false, throwErr = false } = {} as IRunMutateOptions;
   if (typeof descOrOptions === 'string') {
-    return { out, desc: descOrOptions, strict };
+    return { desc: descOrOptions, strict, throwErr };
   }
-  return { out, desc, strict, ...descOrOptions };
+  return { desc, strict, ...descOrOptions, throwErr };
 }
 
 export function parseWatchOptions(options?: WatchOptionsType) {
-  let deps: WatchDepFn = noop;
+  let deps: WatchFnDeps = noop;
   let immediate: boolean = false;
 
   if (isFn(options)) {